requirements.txt
# Core Dependencies
openai>=1.6.1
pyyaml==6.0.1
python-dotenv==1.0.0
httpx>=0.26.0  # For async HTTP requests

# Development Dependencies
pytest==7.4.3
black==23.12.1
flake8==7.0.0